refactor: replace usecase handlers and schema with strategic initiative counterparts
