+++ electron/handlers/strategic-initiative-handler.ts
@@ -2,7 +2,7 @@ import { ipcMain } from "electron";
 import { store } from "../services/store";
 import { AppConfig } from "../schema/core/store.schema";
 import { appendUsecaseRequirement } from "../utils/file";
-import { UsecaseDraftSchema } from "../schema/core/usecase.schema";
+import { StrategicInitiativeDraftSchema } from "../schema/core/strategic-initiative.schema";
 import { generateUseCase } from "../api/core/usecase";
 
 export function setupSIHandlers() {
@@ -10,7 +10,7 @@ export function setupSIHandlers() {
   ipcMain.handle("strategic-initiative:add", async (_, payload) => {
     const APP_CONFIG = store.get<AppConfig>("APP_CONFIG");
     const WORKING_DIR: string = APP_CONFIG?.directoryPath || "";
-    const parsed = await UsecaseDraftSchema.parseAsync(payload);
+    const parsed = await StrategicInitiativeDraftSchema.parseAsync(payload);
     return appendUsecaseRequirement(WORKING_DIR, parsed);
   });
